Updated ports/win32 demo for Visual C++ to compile and run with the latest example objects (BI, BO, and LSP).  Cleaned up a couple of compile warnings, which corrected a bug in the Binary Output object.

+++ trunk/bacnet-stack/bacapp.c
@@ -488,7 +488,7 @@ bool bacapp_parse_application_data(BACNET_APPLICATION_TAG tag_number,
             value->type.Signed_Int = long_value;
         } else if (tag_number == BACNET_APPLICATION_TAG_REAL) {
             double_value = strtod(argv, NULL);
-            value->type.Real = double_value;
+            value->type.Real = (float)double_value;
         } else if (tag_number == BACNET_APPLICATION_TAG_DOUBLE) {
             double_value = strtod(argv, NULL);
             value->type.Double = double_value;
